pyproject.toml
[project]
name = "goldia-dashboard"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "aiocache>=0.12.3",
    "humanize>=4.14.0",
    "motor>=3.7.1",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "streamlit>=1.51.0",
]

[dependency-groups]
dev = [
    "notebook>=7.4.7",
]
